+++ tests/factories/__init__.py
@@ -65,6 +65,7 @@ class HarvesterFactory(DjangoModelFactory):
 
         mock_entry = mock.create_autospec(pkg_resources.EntryPoint, instance=True)
         mock_entry.name = self.key
+        mock_entry.module_name = self.key
         mock_entry.resolve.return_value = MockHarvester
 
         stevedore.ExtensionManager.ENTRY_POINT_CACHE['share.harvesters'].append(mock_entry)